refactor: 회원 조회 기능 ->  MemberService로 이동
