+++ spring-rest-member/src/main/java/com/dife/member/model/dto/RegisterRequestDto.java
@@ -1,4 +1,4 @@
-package com.dife.member.model;
+package com.dife.member.model.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
